tools/requirements.txt
Pillow
pyYAML
